+++ src/hooks/index.js
@@ -11,6 +11,5 @@ export { default as useInventoryBrowse } from './useInventoryBrowse';
 export { default as useLocationsQuery } from './useLocationsQuery';
 export { default as useLastSearchTerms } from './useLastSearchTerms';
 export { default as useSearchForShadowInstanceTenants } from './useSearchForShadowInstanceTenants';
-export { default as useUserTenantPermissions } from './useUserTenantPermissions';
 export { default as useLocationsForTenants } from './useLocationsForTenants';
 export { default as useClassificationIdentifierTypes } from './useClassificationIdentifierTypes';